requirements.txt
# Core NLP & ML
transformers
torch
spacy>=3.5
datasets
seqeval

# API & UI
fastapi
uvicorn
streamlit
requests
pydantic
python-multipart

# LLM Backends
openai
python-dotenv

# PHI Detection
presidio-analyzer
presidio-anonymizer

# OCR
pytesseract
Pillow
pdf2image
